refactor(motor-control): separate stepper motor & brushed motor files (#318)
