+++ head/simulator/main.cpp
@@ -7,9 +7,9 @@
 #include "head/core/presence_sensing_driver.hpp"
 #include "head/core/tasks.hpp"
 #include "head/simulation/adc.hpp"
-#include "motor-control/core/motor.hpp"
-#include "motor-control/core/motor_interrupt_handler.hpp"
-#include "motor-control/core/tmc2130.hpp"
+#include "motor-control/core/stepper_motor/motor.hpp"
+#include "motor-control/core/stepper_motor/motor_interrupt_handler.hpp"
+#include "motor-control/core/stepper_motor/tmc2130.hpp"
 #include "motor-control/simulation/motor_interrupt_driver.hpp"
 #include "motor-control/simulation/sim_motor_hardware_iface.hpp"
 #include "task.h"
