+++ include/motor-control/core/stepper_motor/motor_interrupt_handler.hpp
@@ -2,9 +2,9 @@
 
 #include "common/core/logging.h"
 #include "common/core/message_queue.hpp"
-#include "motor_hardware_interface.hpp"
-#include "motor_messages.hpp"
-#include "tasks/move_status_reporter_task.hpp"
+#include "motor-control/core/motor_hardware_interface.hpp"
+#include "motor-control/core/motor_messages.hpp"
+#include "motor-control/core/tasks/move_status_reporter_task.hpp"
 
 namespace motor_handler {
