+++ include/motor-control/core/tasks/brushed_motion_controller_task.hpp
@@ -6,7 +6,7 @@
 #include "can/core/ids.hpp"
 #include "can/core/messages.hpp"
 #include "common/core/logging.h"
-#include "motor-control/core/motion_controller.hpp"
+#include "motor-control/core/stepper_motor/motion_controller.hpp"
 #include "motor-control/core/tasks/messages.hpp"
 #include "motor-control/core/utils.hpp"
