+++ include/motor-control/firmware/stepper_motor/motor_hardware.hpp
