+++ motor-control/firmware/brushed_motor/brushed_motor_hardware.cpp
@@ -1,6 +1,6 @@
-#include "motor-control/firmware/brushed_motor_hardware.hpp"
+#include "motor-control/firmware/brushed_motor/brushed_motor_hardware.hpp"
 
-#include "motor_control_hardware.h"
+#include "motor-control/firmware/motor_control_hardware.h"
 
 using namespace motor_hardware;
