+++ motor-control/firmware/stepper_motor/motor_hardware.cpp
@@ -1,6 +1,6 @@
-#include "motor-control/firmware/motor_hardware.hpp"
+#include "motor-control/firmware/stepper_motor/motor_hardware.hpp"
 
-#include "motor_control_hardware.h"
+#include "motor-control/firmware/motor_control_hardware.h"
 
 using namespace motor_hardware;
