+++ motor-control/tests/test_motor_driver.cpp
@@ -1,8 +1,8 @@
 #include "catch2/catch.hpp"
 #include "common/simulation/spi.hpp"
-#include "motor-control/core/motor_driver.hpp"
-#include "motor-control/core/motor_driver_config.hpp"
-#include "motor-control/core/tmc2130_registers.hpp"
+#include "motor-control/core/stepper_motor/motor_driver.hpp"
+#include "motor-control/core/stepper_motor/motor_driver_config.hpp"
+#include "motor-control/core/stepper_motor/tmc2130_registers.hpp"
 
 using namespace motor_driver_config;
