+++ motor-control/tests/test_motor_interrupt_queue.cpp
@@ -1,6 +1,6 @@
 #include "catch2/catch.hpp"
 #include "common/tests/mock_message_queue.hpp"
-#include "motor-control/core/motor_interrupt_handler.hpp"
+#include "motor-control/core/stepper_motor/motor_interrupt_handler.hpp"
 #include "motor-control/tests/mock_motor_hardware.hpp"
 #include "motor-control/tests/mock_move_status_reporter_client.hpp"
