+++ pipettes/simulator/main.cpp
@@ -12,8 +12,8 @@
 #include "common/core/logging.h"
 #include "common/simulation/i2c_sim.hpp"
 #include "common/simulation/spi.hpp"
-#include "motor-control/core/motor.hpp"
-#include "motor-control/core/tmc2130.hpp"
+#include "motor-control/core/stepper_motor/motor.hpp"
+#include "motor-control/core/stepper_motor/tmc2130.hpp"
 #include "motor-control/simulation/motor_interrupt_driver.hpp"
 #include "motor-control/simulation/sim_motor_hardware_iface.hpp"
 #include "pipettes/core/configs.hpp"
